+++ src/IconvTranscoder.php
@@ -46,7 +46,7 @@ class IconvTranscoder implements TranscoderInterface
         );
 
         try {
-            $result = iconv($from, $to ?: $this->defaultEncoding, $string);
+            $result = iconv($from ?: '', $to ?: $this->defaultEncoding, $string);
         } finally {
             restore_error_handler();
         }