+++ src/Exception/ExtensionMissingException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Ddeboer\Transcoder\Exception;
 
 class ExtensionMissingException extends \RuntimeException