+++ src/Exception/UndetectableEncodingException.php
@@ -1,14 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Ddeboer\Transcoder\Exception;
 
 class UndetectableEncodingException extends \RuntimeException
 {
-    /**
-     * @param string $string
-     * @param string $error
-     */
-    public function __construct($string, $error)
+    public function __construct(string $string, string $error)
     {
         parent::__construct(sprintf('Encoding for %s is undetectable: %s', $string, $error));
     }