+++ src/MbTranscoder.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Ddeboer\Transcoder;
 
 use Ddeboer\Transcoder\Exception\ExtensionMissingException;
@@ -21,11 +23,9 @@ class MbTranscoder implements TranscoderInterface
     /**
      * Create a Mb-based transcoder.
      *
-     * @param string $defaultEncoding
-     *
      * @throws ExtensionMissingException
      */
-    public function __construct($defaultEncoding = 'UTF-8')
+    public function __construct(string $defaultEncoding = 'UTF-8')
     {
         if (!function_exists('mb_convert_encoding')) {
             throw new ExtensionMissingException('mb');
@@ -44,8 +44,10 @@ class MbTranscoder implements TranscoderInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @param array<string>|string|null $from
      */
-    public function transcode($string, $from = null, $to = null)
+    public function transcode(string $string, $from = null, ?string $to = null): string
     {
         if ($from) {
             if (is_array($from)) {
@@ -62,7 +64,7 @@ class MbTranscoder implements TranscoderInterface
         $handleErrors = !$from || 'auto' === $from;
         if ($handleErrors) {
             set_error_handler(
-                function ($no, $warning) use ($string) {
+                function ($no, $warning) use ($string): void {
                     throw new UndetectableEncodingException($string, $warning);
                 },
                 E_WARNING
@@ -85,27 +87,16 @@ class MbTranscoder implements TranscoderInterface
     }
 
     /**
-     * @param string $encoding
-     * @param bool $allowAuto
-     *
-     * @return void
-     *
      * @throws UnsupportedEncodingException
      */
-    private function assertSupported($encoding, $allowAuto = true)
+    private function assertSupported(string $encoding, bool $allowAuto = true): void
     {
         if (!$this->isSupported($encoding, $allowAuto)) {
             throw new UnsupportedEncodingException($encoding);
         }
     }
 
-    /**
-     * @param string $encoding
-     * @param bool $allowAuto
-     *
-     * @return bool
-     */
-    private function isSupported($encoding, $allowAuto)
+    private function isSupported(string $encoding, bool $allowAuto): bool
     {
         return ($allowAuto && $encoding === 'auto') || isset(self::$encodings[strtolower($encoding)]);
     }