+++ src/TranscoderInterface.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Ddeboer\Transcoder;
 
 use Ddeboer\Transcoder\Exception\UnsupportedEncodingException;
@@ -7,15 +9,12 @@ use Ddeboer\Transcoder\Exception\UnsupportedEncodingException;
 interface TranscoderInterface
 {
     /**
-     * Transcode a string from one into another encoding.
+     * Transcode a string from one encoding into another.
      *
-     * @param string $string String
-     * @param string $from   From encoding (optional)
-     * @param string $to     To encoding (optional)
-     *
-     * @return string
+     * @param string $from Source encoding (optional)
+     * @param string $to Target encoding (optional)
      *
      * @throws UnsupportedEncodingException
      */
-    public function transcode($string, $from = null, $to = null);
+    public function transcode(string $string, ?string $from = null, ?string $to = null): string;
 }