+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## [2.0.0] – 2023-03-07
 
 - Iconv: Fix warning on PHP 8.2 when passing `null` as source encoding.
 - Raise minimum PHP version to 7.2.5.
@@ -23,4 +23,5 @@ The project has been revived and is now available under the name [`fossar/transc
 - Added Nix expression for easier development and sharing the environment with CI.
 - Switched to GitHub Actions for CI and added more PHP versions.
 
+[2.0.0]: https://github.com/fossar/transcoder/compare/1.0.1...v2.0.0
 [1.0.1]: https://github.com/fossar/transcoder/compare/1.0.0...v1.0.1